[closure] Update to latest v20211006 and update removed code
buildFromFile is getting the entry name and fails if not provided.  That is deprecated though so not sure what to replace it with.  Hopefully they will at least show that better when they delete it.  None of the prior told how to even proceed even though most of the internals matched but some of that was deleted too :(